Refactor engine and github packages to reduce PR contention
Split engine/engine.go (1008 lines) into 6 focused files:
- engine.go: Engine struct, constructors, logf, mapKeys
- poll.go: Run, poll, cleanupLockedIssues, checkAndUpgrade, git utilities
- item.go: itemNeedsWork, processItem, extractModelOverride, label helpers
- comments.go: findNewComments, processComments, markCommentsProcessed
- pr.go: ensureDraftPR, ensurePRLinksIssue, markPRReady, postOutputToPR
- stages.go: handleStageComplete, advanceToNextStage

Split github/mutations.go (252 lines) into 4 domain files:
- labels.go: AddLabelToIssue, RemoveLabelFromIssue, ensureLabel
- comments.go: AddComment, AddCommentReaction, UpdateIssueBody, GetIssueBody
- prs.go: CreateDraftPR, MarkPRReady, FindPRForIssue
- status.go: StatusField type, UpdateProjectItemStatus, FetchStatusField

Add adrs/010-file-granularity.md documenting the one-concern-per-file principle.
