+++ engine/stages.go
@@ -1,0 +1,51 @@
+package engine
+
+import (
+	"fmt"
+
+	gh "github.com/verveguy/fabrik/github"
+	"github.com/verveguy/fabrik/stages"
+)
+
+func (e *Engine) handleStageComplete(board *gh.ProjectBoard, item gh.ProjectItem, stage *stages.Stage) {
+	logf(item.Number, "done", "stage %q complete\n", stage.Name)
+
+	completeLabel := fmt.Sprintf("stage:%s:complete", stage.Name)
+	if err := e.client.AddLabelToIssue(e.cfg.Owner, e.cfg.Repo, item.Number, completeLabel); err != nil {
+		logf(item.Number, "warn", "could not add completion label: %v\n", err)
+	}
+
+	shouldAdvance := e.cfg.Yolo
+	if stage.AutoAdvance != nil {
+		shouldAdvance = *stage.AutoAdvance
+	}
+
+	if shouldAdvance {
+		if err := e.advanceToNextStage(board, item, stage); err != nil {
+			logf(item.Number, "warn", "could not advance: %v\n", err)
+		}
+	} else {
+		logf(item.Number, "wait", "waiting for human to advance\n")
+	}
+}
+
+func (e *Engine) advanceToNextStage(board *gh.ProjectBoard, item gh.ProjectItem, currentStage *stages.Stage) error {
+	next := stages.NextStage(e.cfg.Stages, currentStage.Name)
+	if next == nil {
+		logf(item.Number, "info", "completed all stages\n")
+		return nil
+	}
+
+	if e.statusField == nil {
+		return fmt.Errorf("status field metadata not available")
+	}
+
+	optionID, ok := e.statusField.Options[next.Name]
+	if !ok {
+		return fmt.Errorf("no status option %q found on project board (available: %v)",
+			next.Name, mapKeys(e.statusField.Options))
+	}
+
+	logf(item.Number, "advance", "moving to stage %q\n", next.Name)
+	return e.client.UpdateProjectItemStatus(board.ProjectID, item.ItemID, e.statusField.FieldID, optionID)
+}
